pyproject.toml
[project]
name = "kiran-mlops"
version = "0.1.0"
description = "MLOps Final Project - Energy Consumption Prediction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "h2o>=3.44.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "mlflow>=2.9.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "evidently>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
